wholesale changes getting ready for integration tests
