+++ src/tests/integration/lightnet.test.ts
@@ -1,0 +1,19 @@
+import { deploy } from '../../deploy.js';
+import { initBlockchain } from '../../mina.js';
+import { TestHelper } from '../unit/unit-test-helper.js';
+
+import { describe, it, before } from 'node:test';
+
+before(async () => {
+  const chain = await initBlockchain('lightnet');
+  console.log(chain);
+  const deployedContracts = await deploy(chain, chain.keys[0]);
+});
+
+describe('zkUSD Lightnet Test Suite', () => {
+  const testHelper = new TestHelper();
+
+  before(async () => {
+    await testHelper.initChain({ useLightnet: true });
+  });
+});
