+++ src/tests/unit/administration/general-admin.test.ts
@@ -1,8 +1,9 @@
 import { AccountUpdate, Bool, PrivateKey, VerificationKey, Mina } from 'o1js';
-import { TestAmounts, TestHelper } from '../test-helper.js';
-import { ProtocolData } from '../../types.js';
+import { TestAmounts, TestHelper } from '../unit-test-helper.js';
+import { ProtocolData } from '../../../types.js';
 import { describe, it, before } from 'node:test';
 import assert from 'node:assert';
+import { transaction } from '../../../utils/transaction.js';
 
 describe('zkUSD Protocol Vault Administration Test Suite', () => {
   const testHelper = new TestHelper();
@@ -18,7 +19,7 @@ describe('zkUSD Protocol Vault Administration Test Suite', () => {
     await testHelper.createVaults(['alice']);
 
     //Alice deposits 100 Mina
-    await testHelper.transaction(testHelper.agents.alice.keys, async () => {
+    await transaction(testHelper.agents.alice.keys, async () => {
       await testHelper.engine.contract.depositCollateral(
         testHelper.agents.alice.vault!.publicKey,
         TestAmounts.COLLATERAL_100_MINA
@@ -26,20 +27,20 @@ describe('zkUSD Protocol Vault Administration Test Suite', () => {
     });
 
     //Fund the creation of the new admin keys
-    await testHelper.transaction(testHelper.agents.alice.keys, async () => {
+    await transaction(testHelper.agents.alice.keys, async () => {
       AccountUpdate.fundNewAccount(testHelper.agents.alice.keys.publicKey, 1);
       AccountUpdate.create(newAdmin.publicKey);
     });
   });
 
   it('should allow the admin key to be changed with the current admin key', async () => {
-    await testHelper.transaction(
+    await transaction(
       testHelper.agents.alice.keys,
       async () => {
         await testHelper.engine.contract.updateAdmin(newAdmin.publicKey);
       },
       {
-        extraSigners: [TestHelper.protocolAdminKeyPair.privateKey],
+        extraSigners: [testHelper.networkKeys.protocolAdmin.privateKey],
       }
     );
 
@@ -62,12 +63,12 @@ describe('zkUSD Protocol Vault Administration Test Suite', () => {
       // @ts-ignore
 
       latestEvent.event.data.previousAdmin,
-      TestHelper.protocolAdminKeyPair.publicKey
+      testHelper.networkKeys.protocolAdmin.publicKey
     );
   });
 
   it('should allow the new admin key to make updates to the protocol vault', async () => {
-    await testHelper.transaction(
+    await transaction(
       testHelper.agents.alice.keys,
       async () => {
         await testHelper.engine.contract.stopTheProtocol();
@@ -84,7 +85,7 @@ describe('zkUSD Protocol Vault Administration Test Suite', () => {
     assert.deepStrictEqual(protocolData.emergencyStop, Bool(true));
 
     //Resume the protocol
-    await testHelper.transaction(
+    await transaction(
       testHelper.agents.alice.keys,
       async () => {
         await testHelper.engine.contract.resumeTheProtocol();
@@ -100,18 +101,18 @@ describe('zkUSD Protocol Vault Administration Test Suite', () => {
 
   it('should not allow the admin key to be updated without the current admin key', async () => {
     await assert.rejects(async () => {
-      await testHelper.transaction(testHelper.agents.alice.keys, async () => {
+      await transaction(testHelper.agents.alice.keys, async () => {
         await testHelper.engine.contract.updateAdmin(newAdmin.publicKey);
       });
     }, /Transaction verification failed/i);
   });
 
   it('should not allow the admin contract to be upgraded in the current version', async () => {
-    const oldAccount = Mina.getAccount(testHelper.engine.publicKey);
+    const oldAccount = Mina.getAccount(testHelper.networkKeys.engine.publicKey);
     const verificationKey = oldAccount.zkapp?.verificationKey;
 
     await assert.rejects(async () => {
-      await testHelper.transaction(
+      await transaction(
         testHelper.deployer,
         async () => {
           await testHelper.engine.contract.updateVerificationKey(
