+++ src/tests/unit/administration/oracle-fee.test.ts
@@ -1,8 +1,9 @@
 import { AccountUpdate, Field, Mina, PrivateKey, UInt32 } from 'o1js';
-import { TestAmounts, TestHelper } from '../test-helper.js';
-import { ProtocolData } from '../../types.js';
+import { TestAmounts, TestHelper } from '../unit-test-helper.js';
+import { ProtocolData } from '../../../types.js';
 import { describe, it, before } from 'node:test';
 import assert from 'node:assert';
+import { transaction } from '../../../utils/transaction.js';
 
 describe('zkUSD Protocol Oracle Fee Test Suite', () => {
   const testHelper = new TestHelper();
@@ -16,13 +17,13 @@ describe('zkUSD Protocol Oracle Fee Test Suite', () => {
   it('should allow the fee to be changed with the admin key', async () => {
     const newFee = TestAmounts.COLLATERAL_2_MINA;
 
-    await testHelper.transaction(
+    await transaction(
       testHelper.deployer,
       async () => {
         await testHelper.engine.contract.updateOracleFee(newFee);
       },
       {
-        extraSigners: [TestHelper.protocolAdminKeyPair.privateKey],
+        extraSigners: [testHelper.networkKeys.protocolAdmin.privateKey],
       }
     );
 
@@ -56,27 +57,31 @@ describe('zkUSD Protocol Oracle Fee Test Suite', () => {
     const newFee = TestAmounts.COLLATERAL_1_MINA;
 
     await assert.rejects(async () => {
-      await testHelper.transaction(testHelper.agents.alice.keys, async () => {
+      await transaction(testHelper.agents.alice.keys, async () => {
         await testHelper.engine.contract.updateOracleFee(newFee);
       });
     }, /Transaction verification failed/i);
   });
 
   it('should not allow the private key to manually send funds from the engine', async () => {
-    const oracleBalanceBefore = Mina.getBalance(testHelper.engine.publicKey);
+    const oracleBalanceBefore = Mina.getBalance(
+      testHelper.networkKeys.engine.publicKey
+    );
 
     await assert.rejects(async () => {
-      await testHelper.transaction(
+      await transaction(
         testHelper.agents.alice.keys,
         async () => {
-          const sendUpdate = AccountUpdate.create(testHelper.engine.publicKey);
+          const sendUpdate = AccountUpdate.create(
+            testHelper.networkKeys.engine.publicKey
+          );
           sendUpdate.send({
             to: testHelper.agents.alice.keys.publicKey,
             amount: oracleBalanceBefore,
           });
         },
         {
-          extraSigners: [TestHelper.engineKeyPair.privateKey],
+          extraSigners: [testHelper.networkKeys.engine.privateKey],
         }
       );
     }, /Update_not_permitted_balance/i);
@@ -90,7 +95,7 @@ describe('zkUSD Protocol Oracle Fee Test Suite', () => {
 
     //get the current balance of the price feed oracle
     const priceFeedOracleBalanceBefore = Mina.getBalance(
-      testHelper.engine.publicKey
+      testHelper.networkKeys.engine.publicKey
     );
     const whitelistedOracles = testHelper.whitelistedOracles;
     const oracleName = Array.from(whitelistedOracles.keys())[0];
@@ -100,7 +105,7 @@ describe('zkUSD Protocol Oracle Fee Test Suite', () => {
     const oracleBalanceBefore = Mina.getBalance(oracle.publicKey);
 
     // Submit price from oracle
-    await testHelper.transaction(
+    await transaction(
       oracle,
       async () => {
         await testHelper.engine.contract.submitPrice(
@@ -117,7 +122,7 @@ describe('zkUSD Protocol Oracle Fee Test Suite', () => {
     const oracleBalanceAfter = Mina.getBalance(oracle.publicKey);
 
     const priceFeedOracleBalanceAfter = Mina.getBalance(
-      testHelper.engine.publicKey
+      testHelper.networkKeys.engine.publicKey
     );
 
     // Verify oracle received the fee
