+++ src/tests/unit/price-feed/price-submission.test.ts
@@ -7,21 +7,22 @@ import {
   UInt32,
   UInt64,
 } from 'o1js';
-import { TestAmounts, TestHelper } from '../test-helper.js';
+import { TestAmounts, TestHelper } from '../unit-test-helper.js';
 import {
   OracleWhitelist,
   PriceSubmission,
   PriceSubmissionPacked,
   ProtocolData,
-} from '../../types.js';
-import { ZkUsdEngine, ZkUsdEngineErrors } from '../../zkusd-engine.js';
+} from '../../../types.js';
+import { ZkUsdEngineErrors } from '../../../contracts/zkusd-engine.js';
 import { describe, it, before, beforeEach } from 'node:test';
 import assert from 'node:assert';
 import {
   ZkUsdMasterOracle,
   ZkUsdMasterOracleErrors,
-} from '../../zkusd-master-oracle.js';
-import { ZkUsdPriceTracker } from '../../zkusd-price-tracker.js';
+} from '../../../contracts/zkusd-master-oracle.js';
+import { ZkUsdPriceTracker } from '../../../contracts/zkusd-price-tracker.js';
+import { transaction } from '../../../utils/transaction.js';
 
 describe('zkUSD Price Feed Oracle Submission Test Suite', () => {
   const testHelper = new TestHelper();
@@ -45,8 +46,8 @@ describe('zkUSD Price Feed Oracle Submission Test Suite', () => {
       .toBoolean();
 
     return isEven
-      ? ZkUsdEngine.EVEN_ORACLE_PRICE_TRACKER_ADDRESS
-      : ZkUsdEngine.ODD_ORACLE_PRICE_TRACKER_ADDRESS;
+      ? testHelper.networkKeys.evenOraclePriceTracker.publicKey
+      : testHelper.networkKeys.oddOraclePriceTracker.publicKey;
   };
 
   beforeEach(async () => {
@@ -57,7 +58,7 @@ describe('zkUSD Price Feed Oracle Submission Test Suite', () => {
     };
     testHelper.whitelistedOracles = new Map(whitelistedOracles);
 
-    await testHelper.transaction(
+    await transaction(
       testHelper.agents.alice.keys,
       async () => {
         await testHelper.engine.contract.updateOracleWhitelist(
@@ -65,12 +66,12 @@ describe('zkUSD Price Feed Oracle Submission Test Suite', () => {
         );
       },
       {
-        extraSigners: [TestHelper.protocolAdminKeyPair.privateKey],
+        extraSigners: [testHelper.networkKeys.protocolAdmin.privateKey],
       }
     );
 
     //settle any outstanding actions
-    await testHelper.transaction(testHelper.agents.alice.keys, async () => {
+    await transaction(testHelper.agents.alice.keys, async () => {
       await testHelper.engine.contract.settlePriceUpdate();
     });
   });
@@ -79,7 +80,7 @@ describe('zkUSD Price Feed Oracle Submission Test Suite', () => {
     const whitelistedOracles = testHelper.whitelistedOracles;
     const oracleName = Array.from(whitelistedOracles.keys())[0];
 
-    await testHelper.transaction(testHelper.oracles[oracleName], async () => {
+    await transaction(testHelper.oracles[oracleName], async () => {
       await testHelper.engine.contract.submitPrice(
         TestAmounts.PRICE_48_CENT,
         testHelper.whitelist
@@ -105,7 +106,7 @@ describe('zkUSD Price Feed Oracle Submission Test Suite', () => {
     const whitelistedOracles = testHelper.whitelistedOracles;
     const oracleName = Array.from(whitelistedOracles.keys())[0];
 
-    await testHelper.transaction(testHelper.oracles[oracleName], async () => {
+    await transaction(testHelper.oracles[oracleName], async () => {
       await testHelper.engine.contract.submitPrice(
         TestAmounts.PRICE_52_CENT,
         testHelper.whitelist
@@ -116,7 +117,7 @@ describe('zkUSD Price Feed Oracle Submission Test Suite', () => {
 
     assert.deepStrictEqual(
       trackerAddress,
-      ZkUsdEngine.EVEN_ORACLE_PRICE_TRACKER_ADDRESS
+      testHelper.networkKeys.evenOraclePriceTracker.publicKey
     );
 
     const tracker = new ZkUsdPriceTracker(
@@ -136,7 +137,7 @@ describe('zkUSD Price Feed Oracle Submission Test Suite', () => {
     const whitelistedOracles = testHelper.whitelistedOracles;
     const oracleName = Array.from(whitelistedOracles.keys())[0];
 
-    await testHelper.transaction(testHelper.oracles[oracleName], async () => {
+    await transaction(testHelper.oracles[oracleName], async () => {
       await testHelper.engine.contract.submitPrice(
         TestAmounts.PRICE_49_CENT,
         testHelper.whitelist
@@ -147,7 +148,7 @@ describe('zkUSD Price Feed Oracle Submission Test Suite', () => {
 
     assert.deepStrictEqual(
       trackerAddress,
-      ZkUsdEngine.ODD_ORACLE_PRICE_TRACKER_ADDRESS
+      testHelper.networkKeys.oddOraclePriceTracker.publicKey
     );
 
     const tracker = new ZkUsdPriceTracker(
@@ -165,7 +166,7 @@ describe('zkUSD Price Feed Oracle Submission Test Suite', () => {
     const whitelistedOracles = testHelper.whitelistedOracles;
     const oracleName = Array.from(whitelistedOracles.keys())[0];
 
-    await testHelper.transaction(testHelper.oracles[oracleName], async () => {
+    await transaction(testHelper.oracles[oracleName], async () => {
       await testHelper.engine.contract.submitPrice(
         TestAmounts.PRICE_25_CENT,
         testHelper.whitelist
@@ -195,7 +196,7 @@ describe('zkUSD Price Feed Oracle Submission Test Suite', () => {
 
   it('should not allow a non-whitelisted address to submit a price', async () => {
     await assert.rejects(async () => {
-      await testHelper.transaction(testHelper.agents.alice.keys, async () => {
+      await transaction(testHelper.agents.alice.keys, async () => {
         await testHelper.engine.contract.submitPrice(
           TestAmounts.PRICE_25_CENT,
           testHelper.whitelist
@@ -209,7 +210,7 @@ describe('zkUSD Price Feed Oracle Submission Test Suite', () => {
     fakeWhitelist.addresses[0] = testHelper.agents.alice.keys.publicKey;
 
     await assert.rejects(async () => {
-      await testHelper.transaction(testHelper.agents.alice.keys, async () => {
+      await transaction(testHelper.agents.alice.keys, async () => {
         await testHelper.engine.contract.submitPrice(
           TestAmounts.PRICE_25_CENT,
           fakeWhitelist
@@ -223,7 +224,7 @@ describe('zkUSD Price Feed Oracle Submission Test Suite', () => {
     const oracleName = Array.from(whitelistedOracles.keys())[0];
 
     await assert.rejects(async () => {
-      await testHelper.transaction(testHelper.oracles[oracleName], async () => {
+      await transaction(testHelper.oracles[oracleName], async () => {
         await testHelper.engine.contract.submitPrice(
           TestAmounts.ZERO,
           testHelper.whitelist
@@ -237,7 +238,7 @@ describe('zkUSD Price Feed Oracle Submission Test Suite', () => {
     const oracleNames = Array.from(whitelistedOracles.keys());
 
     for (const oracleName of oracleNames) {
-      await testHelper.transaction(testHelper.oracles[oracleName], async () => {
+      await transaction(testHelper.oracles[oracleName], async () => {
         await testHelper.engine.contract.submitPrice(
           TestAmounts.PRICE_25_CENT,
           testHelper.whitelist
@@ -316,14 +317,15 @@ describe('zkUSD Price Feed Oracle Submission Test Suite', () => {
     const whitelistedOracles = testHelper.whitelistedOracles;
     const oracleName = Array.from(whitelistedOracles.keys())[0];
 
-    await testHelper.transaction(testHelper.oracles[oracleName], async () => {
+    await transaction(testHelper.oracles[oracleName], async () => {
       await testHelper.engine.contract.submitPrice(
         TestAmounts.PRICE_2_USD,
         testHelper.whitelist
       );
     });
 
-    const trackerAddress = ZkUsdEngine.ODD_ORACLE_PRICE_TRACKER_ADDRESS;
+    const trackerAddress =
+      testHelper.networkKeys.oddOraclePriceTracker.publicKey;
 
     const tracker = new ZkUsdPriceTracker(
       trackerAddress,
@@ -337,7 +339,7 @@ describe('zkUSD Price Feed Oracle Submission Test Suite', () => {
   });
 
   it('should allow the fallback price to be updated with the admin key', async () => {
-    await testHelper.transaction(
+    await transaction(
       testHelper.deployer,
       async () => {
         await testHelper.engine.contract.updateFallbackPrice(
@@ -345,7 +347,7 @@ describe('zkUSD Price Feed Oracle Submission Test Suite', () => {
         );
       },
       {
-        extraSigners: [TestHelper.protocolAdminKeyPair.privateKey],
+        extraSigners: [testHelper.networkKeys.protocolAdmin.privateKey],
       }
     );
 
@@ -364,7 +366,7 @@ describe('zkUSD Price Feed Oracle Submission Test Suite', () => {
   });
 
   it('should emit the fallback price update event', async () => {
-    await testHelper.transaction(
+    await transaction(
       testHelper.deployer,
       async () => {
         await testHelper.engine.contract.updateFallbackPrice(
@@ -372,7 +374,7 @@ describe('zkUSD Price Feed Oracle Submission Test Suite', () => {
         );
       },
       {
-        extraSigners: [TestHelper.protocolAdminKeyPair.privateKey],
+        extraSigners: [testHelper.networkKeys.protocolAdmin.privateKey],
       }
     );
 
@@ -389,7 +391,7 @@ describe('zkUSD Price Feed Oracle Submission Test Suite', () => {
 
   it('should update the even fallback price if we are on an odd block', async () => {
     testHelper.chain.local?.setBlockchainLength(UInt32.from(1));
-    await testHelper.transaction(
+    await transaction(
       testHelper.deployer,
       async () => {
         await testHelper.engine.contract.updateFallbackPrice(
@@ -397,7 +399,7 @@ describe('zkUSD Price Feed Oracle Submission Test Suite', () => {
         );
       },
       {
-        extraSigners: [TestHelper.protocolAdminKeyPair.privateKey],
+        extraSigners: [testHelper.networkKeys.protocolAdmin.privateKey],
       }
     );
 
@@ -412,7 +414,7 @@ describe('zkUSD Price Feed Oracle Submission Test Suite', () => {
 
   it('should update the odd fallback price if we are on an even block', async () => {
     testHelper.chain.local?.setBlockchainLength(UInt32.from(2));
-    await testHelper.transaction(
+    await transaction(
       testHelper.deployer,
       async () => {
         await testHelper.engine.contract.updateFallbackPrice(
@@ -420,7 +422,7 @@ describe('zkUSD Price Feed Oracle Submission Test Suite', () => {
         );
       },
       {
-        extraSigners: [TestHelper.protocolAdminKeyPair.privateKey],
+        extraSigners: [testHelper.networkKeys.protocolAdmin.privateKey],
       }
     );
 
@@ -435,7 +437,7 @@ describe('zkUSD Price Feed Oracle Submission Test Suite', () => {
 
   it('should not allow the fallback price to be updated without the admin key', async () => {
     await assert.rejects(async () => {
-      await testHelper.transaction(testHelper.deployer, async () => {
+      await transaction(testHelper.deployer, async () => {
         await testHelper.engine.contract.updateFallbackPrice(
           TestAmounts.PRICE_2_USD
         );
@@ -445,7 +447,7 @@ describe('zkUSD Price Feed Oracle Submission Test Suite', () => {
 
   it('should not allow the fallback price to be updated to 0', async () => {
     await assert.rejects(async () => {
-      await testHelper.transaction(
+      await transaction(
         testHelper.deployer,
         async () => {
           await testHelper.engine.contract.updateFallbackPrice(
@@ -453,7 +455,7 @@ describe('zkUSD Price Feed Oracle Submission Test Suite', () => {
           );
         },
         {
-          extraSigners: [TestHelper.protocolAdminKeyPair.privateKey],
+          extraSigners: [testHelper.networkKeys.protocolAdmin.privateKey],
         }
       );
     }, new RegExp(ZkUsdMasterOracleErrors.AMOUNT_ZERO));
@@ -467,7 +469,7 @@ describe('zkUSD Price Feed Oracle Submission Test Suite', () => {
 
     //get the current balance of the price feed oracle
     const minaBalanceOfEngineBefore = Mina.getBalance(
-      testHelper.engine.publicKey
+      testHelper.networkKeys.engine.publicKey
     );
 
     const oracleFundsInEngineBefore =
@@ -481,7 +483,7 @@ describe('zkUSD Price Feed Oracle Submission Test Suite', () => {
     const oracleBalanceBefore = Mina.getBalance(oracle.publicKey);
 
     // Submit price from oracle
-    await testHelper.transaction(oracle, async () => {
+    await transaction(oracle, async () => {
       await testHelper.engine.contract.submitPrice(
         TestAmounts.PRICE_25_CENT,
         testHelper.whitelist
@@ -492,7 +494,7 @@ describe('zkUSD Price Feed Oracle Submission Test Suite', () => {
     const oracleBalanceAfter = Mina.getBalance(oracle.publicKey);
 
     const priceFeedOracleBalanceAfter = Mina.getBalance(
-      testHelper.engine.publicKey
+      testHelper.networkKeys.engine.publicKey
     );
 
     const oracleFundsInEngineAfter =
@@ -528,7 +530,7 @@ describe('zkUSD Price Feed Oracle Submission Test Suite', () => {
       await testHelper.engine.contract.getAvailableOracleFunds();
 
     //set the fee to the balance of the oracle to drain the funds
-    await testHelper.transaction(
+    await transaction(
       testHelper.deployer,
       async () => {
         await testHelper.engine.contract.updateOracleFee(
@@ -536,7 +538,7 @@ describe('zkUSD Price Feed Oracle Submission Test Suite', () => {
         );
       },
       {
-        extraSigners: [TestHelper.protocolAdminKeyPair.privateKey],
+        extraSigners: [testHelper.networkKeys.protocolAdmin.privateKey],
       }
     );
 
@@ -544,7 +546,7 @@ describe('zkUSD Price Feed Oracle Submission Test Suite', () => {
     const whitelistedOracles = testHelper.whitelistedOracles;
     const oracleName = Array.from(whitelistedOracles.keys())[0];
 
-    await testHelper.transaction(testHelper.oracles[oracleName], async () => {
+    await transaction(testHelper.oracles[oracleName], async () => {
       await testHelper.engine.contract.submitPrice(
         TestAmounts.PRICE_25_CENT,
         testHelper.whitelist
@@ -559,7 +561,7 @@ describe('zkUSD Price Feed Oracle Submission Test Suite', () => {
       UInt64.zero.toString()
     );
 
-    await testHelper.transaction(
+    await transaction(
       testHelper.deployer,
       async () => {
         await testHelper.engine.contract.updateOracleFee(
@@ -567,13 +569,13 @@ describe('zkUSD Price Feed Oracle Submission Test Suite', () => {
         );
       },
       {
-        extraSigners: [TestHelper.protocolAdminKeyPair.privateKey],
+        extraSigners: [testHelper.networkKeys.protocolAdmin.privateKey],
       }
     );
 
     //Submit a new price - this should fail
     await assert.rejects(async () => {
-      await testHelper.transaction(testHelper.oracles[oracleName], async () => {
+      await transaction(testHelper.oracles[oracleName], async () => {
         await testHelper.engine.contract.submitPrice(
           TestAmounts.PRICE_50_CENT,
           testHelper.whitelist
@@ -590,7 +592,7 @@ describe('zkUSD Price Feed Oracle Submission Test Suite', () => {
     );
 
     await assert.rejects(async () => {
-      await testHelper.transaction(testHelper.deployer, async () => {
+      await transaction(testHelper.deployer, async () => {
         const accountUpdate = AccountUpdate.create(
           trackerAddress,
           testHelper.engine.contract.deriveTokenId()
